Remove unneeded scripting manager and add in Python error logging to PythonScript
