+++ src/swganh/scripting/python_script.cc
@@ -51,7 +51,7 @@ void PythonScript::Run()
     }
     catch (error_already_set &)
     {
-        PyErr_Print();
+        GetPythonException();
     }
 }
 
@@ -63,4 +63,54 @@ void PythonScript::ReadFileContents()
     filecontents_ = string(
         (istreambuf_iterator<char>(filestream)),
         istreambuf_iterator<char>());
+}
+
+void PythonScript::GetPythonException()
+{
+    std::ostringstream os;
+    os << "Python error:\n  " << std::flush;
+
+    PyObject *type = 0, *val = 0, *tb = 0;
+    PyErr_Fetch(&type, &val, &tb);
+    handle<> e_val(val), e_type(type), e_tb(allow_null(tb));
+
+    try {
+        object t = extract<object>(e_type.get());
+        object t_name = t.attr("__name__");
+        std::string typestr = extract<std::string>(t_name);
+
+        os << typestr << std::flush;
+    } catch (error_already_set const &) {
+        os << "Internal error getting error type:\n";
+        PyErr_Print();
+    }
+
+    os << ": ";
+
+    try {
+        object v = extract<object>(e_val.get());
+        std::string valuestr = extract<std::string>(v.attr("__str__")());
+        os  << valuestr << std::flush;
+    } catch (error_already_set const &) {
+        os << "Internal error getting value type:\n";
+        PyErr_Print();
+    }
+
+    if (tb) {
+        try {
+            object tb_list = import("traceback").attr("format_tb")(e_tb);
+            object tb_str = str("").attr("join")(tb_list);
+            std::string str = extract<std::string>(tb_str);
+
+            os << "\nTraceback (recent call last):\n" << str;
+        } catch (error_already_set const &) {
+            os << "Internal error getting traceback:\n";
+            PyErr_Print();
+        }
+    } else {
+        os << std::endl;
+    }
+    PyErr_Clear();
+    std::cerr << os.str() << endl;
+    BOOST_LOG_TRIVIAL(warning) << os.str();
 }
